# src/main/openapi/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/expenses-travel/{id}": {
      "put": {
        "tags": ["travel-expense-controller"],
        "operationId": "putTravelExpense",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelExpenseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses-cost/{id}": {
      "put": {
        "tags": ["cost-expense-controller"],
        "operationId": "putCostExpense",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostExpenseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/expenses/files": {
      "post": {
        "tags": ["expense-controller"],
        "operationId": "postFiles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": ["file"],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses-travel": {
      "post": {
        "tags": ["travel-expense-controller"],
        "operationId": "postTravelExpense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelExpenseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses-cost": {
      "post": {
        "tags": ["cost-expense-controller"],
        "operationId": "postCostExpense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostExpenseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/expenses": {
      "get": {
        "tags": ["expense-controller"],
        "operationId": "getExpenseAll",
        "parameters": [
          {
            "name": "personId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/{id}": {
      "get": {
        "tags": ["expense-controller"],
        "operationId": "getExpenseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["expense-controller"],
        "operationId": "deleteExpenseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/files/{file}/{name}": {
      "get": {
        "tags": ["expense-controller"],
        "operationId": "getFiles",
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "byte"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TravelExpenseInput": {
        "type": "object",
        "required": ["personId", "description", "date", "status"],
        "properties": {
          "personId": {
            "$ref": "#/components/schemas/UUID"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["REQUESTED", "APPROVED", "REJECTED", "DONE"]
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "allowance": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "UUID": {
        "type": "object",
        "required": ["value"],
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
      "CostExpenseDetails": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostExpenseFile"
            }
          }
        }
      },
      "CostExpenseFile": {
        "type": "object",
        "required": ["name", "file"],
        "properties": {
          "name": {
            "type": "string"
          },
          "file": {
            "$ref": "#/components/schemas/UUID"
          }
        }
      },
      "Expense": {
        "type": "object",
        "required": ["id", "personId"],
        "properties": {
          "id": {
            "type": "string"
          },
          "personId": {
            "$ref": "#/components/schemas/UUID"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["REQUESTED", "APPROVED", "REJECTED", "DONE"]
          },
          "expenseType": {
            "type": "string",
            "enum": ["TRAVEL", "COST"]
          },
          "costDetails": {
            "$ref": "#/components/schemas/CostExpenseDetails"
          },
          "travelDetails": {
            "$ref": "#/components/schemas/TravelExpenseDetails"
          }
        }
      },
      "TravelExpenseDetails": {
        "type": "object",
        "properties": {
          "distance": {
            "type": "number",
            "format": "double"
          },
          "allowance": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CostExpenseFileInput": {
        "type": "object",
        "required": ["name", "file"],
        "properties": {
          "name": {
            "type": "string"
          },
          "file": {
            "$ref": "#/components/schemas/UUID"
          }
        }
      },
      "CostExpenseInput": {
        "type": "object",
        "required": ["personId", "status", "files"],
        "properties": {
          "personId": {
            "$ref": "#/components/schemas/UUID"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["REQUESTED", "APPROVED", "REJECTED", "DONE"]
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostExpenseFileInput"
            }
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}